Extract some utility code from `manage/Realm/index.tsx` into new file
This breaks up two circular imports
