+++ frontend/src/routes/manage/Realm/AddChild.tsx
@@ -14,7 +14,8 @@ import { useForm } from "react-hook-form";
 import { Input } from "../../../ui/Input";
 import { Form } from "../../../ui/Form";
 import { PathSegmentInput } from "../../../ui/PathSegmentInput";
-import { ErrorBox, NoPath, PathInvalid, realmValidations } from ".";
+import { NoPath, PathInvalid } from ".";
+import { ErrorBox, realmValidations } from "./util";
 import { Button } from "../../../ui/Button";
 import { AddChildMutationResponse } from "../../../query-types/AddChildMutation.graphql";
 import { Spinner } from "../../../ui/Spinner";
